Install a specific version of npm in the Rust builds
Currently, we're just using the version of npm that comes with Node.js
versions in order to avoid what I found to be a pretty annoying case on
Windows where npm couldn't be upgraded because it couldn't replace itself
because it was in use.

So I had an idea to use yarn, which is actually the same way npm had to be
upgraded to 6 for a while because of a bug in npm 5 that shipped with common
Node.js versions.